BACKEND/requirements.txt
# Django Backend Requirements
# Data Masking and Anonymization Tool - Phase 4

# Core Django Framework
django>=4.2,<5.0

# Django REST Framework for API development
djangorestframework>=3.14,<4.0

# JWT Authentication for DRF
djangorestframework-simplejwt>=5.3,<6.0

# CORS headers for frontend integration
django-cors-headers>=4.3,<5.0

# Python dotenv for environment variables (future use)
python-dotenv>=1.0,<2.0

# Database Drivers for Real Database Connections
# PostgreSQL driver
psycopg2-binary>=2.9,<3.0

# MySQL driver
mysql-connector-python>=8.0,<9.0

# MongoDB driver
pymongo>=4.0,<5.0
